Move init inside commands.
