+++ hatch/commands/__init__.py
@@ -2,3 +2,4 @@ from hatch.commands.conda import conda
 from hatch.commands.python import python
 from hatch.commands.config import config
 from hatch.commands.new import new
+from hatch.commands.init import init
